+++ lwk_bindings/tests/bindings/simplicity_p2pk_regtest.py
@@ -14,7 +14,7 @@ client = ElectrumClient.from_url(node.electrum_url())
 # 2. Create signer and derive x-only public key
 mnemonic = Mnemonic("abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon about")
 signer = Signer(mnemonic, network)
-derivation_path = "m/86'/1'/0'/0/0"
+derivation_path = DerivationPath("m/86'/1'/0'/0/0")
 xonly_pubkey = simplicity_derive_xonly_pubkey(signer, derivation_path)
 
 # 3. Compile P2PK program with the public key