+++ lwk_wasm/src/signer.rs
@@ -1,6 +1,6 @@
 use std::collections::HashMap;
 
-use crate::{Bip, Error, Mnemonic, Network, Pset, WolletDescriptor, Xpub};
+use crate::{Bip, DerivationPath, Error, Mnemonic, Network, Pset, WolletDescriptor, Xpub};
 use lwk_common::Signer as SignerTrait;
 use lwk_wollet::{
     bitcoin::bip32, bitcoin::sign_message::MessageSignature,
@@ -75,6 +75,15 @@ impl Signer {
     /// Return the signer fingerprint
     pub fn fingerprint(&self) -> Result<String, Error> {
         Ok(self.inner.fingerprint().to_string())
+    }
+
+    /// Derive an xpub at `path` and return it as a keyorigin xpub string
+    #[wasm_bindgen(js_name = keyoriginXpubFromPath)]
+    pub fn keyorigin_xpub_from_path(&self, path: &DerivationPath) -> Result<String, Error> {
+        let path: bip32::DerivationPath = path.into();
+        let fingerprint = self.inner.fingerprint();
+        let xpub = lwk_common::Signer::derive_xpub(&self.inner, &path)?;
+        Ok(format!("[{fingerprint}/{path}]{xpub}"))
     }
 
     /// Return the mnemonic of the signer