+++ lwk_bindings/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Added `DerivationPath`
+* Removed `get_path()`, replaced with `DerivationPath::from_account()`
+* Changed "simplicity" functions `derive_keypair()`, `simplicity_derive_xonly_pubkey()`, `create_p2pk_signature()` to accept `DerivatioPath` instead of `str`
+
 ## 0.18.3
 
 ## 0.18.2